Add maximum lifetime to cache entries
The cache now supports a maximum lifetime for its entries. The
deletion of expired entries will happen in a lazy way. That means
that expired entries are not removed automatically at the time of
expiration but they removed every time someone is trying to get
this entry through get() or check if the entry is there through
has() method.